cat version

+++ HelloWorld.java
@@ -1,7 +1,7 @@
 public class HelloWorld {
 
     public static void main(String[] args) {
-		private double deffense;
+		private float deffense;
 
         System.out.println("Hello World!");
     }
